+++ tests/assets/testpackage/testpackage/__init__.py
@@ -1,0 +1,1 @@
+"""A test package for use in import-linter's test suite."""