+++ tests/assets/testpackage/testpackage/indirect/__init__.py
@@ -1,0 +1,1 @@
+"""Modules with indirect dependencies between layers."""